refactor: BetAmount 를 game 패키지로 이동
