+++ src/main/java/blackjack/domain/game/BetAmount.java
@@ -1,4 +1,4 @@
-package blackjack.domain.result;
+package blackjack.domain.game;
 
 public record BetAmount(int amount) {
 }
