+++ src/test/java/blackjack/domain/game/PlayersTest.java
@@ -1,6 +1,5 @@
 package blackjack.domain.game;
 
-import blackjack.domain.result.BetAmount;
 import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
