+++ src/test/java/blackjack/domain/profit/DealerProfitsTest.java
@@ -2,7 +2,7 @@ package blackjack.domain.profit;
 
 import blackjack.domain.game.Hand;
 import blackjack.domain.game.Player;
-import blackjack.domain.result.BetAmount;
+import blackjack.domain.game.BetAmount;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
