+++ src/test/java/blackjack/domain/result/JudgeTest.java
@@ -3,6 +3,7 @@ package blackjack.domain.result;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardRank;
 import blackjack.domain.card.CardSuit;
+import blackjack.domain.game.BetAmount;
 import blackjack.domain.game.Dealer;
 import blackjack.domain.game.Hand;
 import blackjack.domain.game.Player;
